backend/requirements.txt
aio-pika==6.8.0
Django==2.2.24  # https://github.com/owkin/connect-backend/issues/121
celery==5.1.2
checksumdir==1.2.0
cryptography==3.4.7
django-cors-headers==3.7.0
django-celery-results==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-spectacular==0.17.3
drf-nested-routers==0.93.3
pyjwt==2.1.0
ipython==7.25.0
ipython-genutils==0.2.0
kubernetes==17.17.0
psycopg2-binary==2.8.6  # psycopg 2.9+ is not compatible with Django <3.2
requests>=2.20.0
tldextract==3.1.0
uwsgi==2.0.19.1
zxcvbn==4.4.28
boto3==1.11.9
grpcio==1.39.0
grpcio-tools==1.36.1
# dev-mode
watchdog==2.1.3
argh==0.26.2
# tests
coverage==5.5
mock==4.0.3
parameterized==0.8.1
